Refactor cell inheritance hierarchy
